requirements.txt
selenium==3.5.0
PyYAML==3.12
peewee==2.10.1
